pipes_game: Implement drawing of the grid to the image
This commit changes the main app to use the CLI-provided filepath and
convert it into a grid. The grid is then sent into new code which draws
the grid onto a frame. Thus, when the app is run with a valid grid text
file, the grid is displayed in the window.