grid: Track pipe endpoints
In order to solve the grid, we need to track the endpoints of each pipe,
and analyse their state. This commit adds the storage for the data, and
all the methods required to initialise it and query it. `set_cell` is
updated to modify the endpoint state, thus allowing us to start to
complete the pipes.